Add JSON contract, use in ng app and go test
A contract in JSON states a method, path, and response object. The
response object is an example response, which the go test will use to
match against the types (not the strict values).

As long as the consumer (ng app) is using the contract to make the
request, then the producer (go server) will be testing against the
actual request being made.