Use make_types to generate interface from contract
Use the following shellscript to generate types from the contract JSON
file, then use the type of the `response` example object to define the
`http.get` response type.
```
$(npm bin)/make_types -i src/app/contract/get-hello.d.ts \
    src/app/contract/get-hello.json HelloResponse
```

Now a contract is fully defined between the producer and consumer, and
it is ensured to be updated by being fully used by the consumer to make
the actual request.